python/requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# Control plane — agent management (Entra ID required)
# ─────────────────────────────────────────────────────────────────────────────
azure-ai-projects>=1.0.0b9

# ─────────────────────────────────────────────────────────────────────────────
# Runtime plane — Option A: Azure AI Foundry inference endpoint (API key)
# ─────────────────────────────────────────────────────────────────────────────
azure-ai-inference>=1.0.0b8

# ─────────────────────────────────────────────────────────────────────────────
# Runtime plane — Option B: openai SDK with Azure OpenAI endpoint (API key)
# ─────────────────────────────────────────────────────────────────────────────
openai>=1.50.0

# ─────────────────────────────────────────────────────────────────────────────
# Entra ID authentication
# ─────────────────────────────────────────────────────────────────────────────
azure-identity>=1.19.0

# ─────────────────────────────────────────────────────────────────────────────
# Optional: load .env during local development
# ─────────────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
